+++ CHANGELOG.rst
@@ -1,3 +1,7 @@
+Unreleased
+----------
+* Updated dataset
+
 v0.56.0
 -------
 * BREAKING CHANGE: ``cs`` column in ``market_watch_init`` and ``market_watch_plus`` now is of string type instead of numeric.